Move routes into own modules
