+++ lobby-server/src/routes/mod.rs
@@ -1,0 +1,3 @@
+pub mod create;
+pub mod join;
+pub mod list;
